requirements.txt
flask == 3.1.2
bcrypt == 5.0.0
python-dotenv == 1.2.1
pyjwt == 2.10.1